feat: 交換ページの分離
